:sparkles: create script to generate fake FASTA from VCF
Adds a script that generates a fake FASTA file from a VCF file.
This is useful when a VCF file lacks contig information in the header,
allowing tools like bcftools reheader to add contig lines based on
the generated FASTA. The script fills the FASTA with 'N's and
replaces them with REF alleles from the VCF, where available.