:sparkles: add sample annotation functionality
Adds a new command to annotate tree sequences with sample metadata
(breed and sample ID). This includes population and individual
metadata, which are then added to the tree sequence tables,
enabling richer analysis and interpretation of the data. The
functionality relies on a sample file to map individual samples to
their breed.  Also, the module version is updated to `0.5.2`.